requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-cohere
langchain-huggingface
rank_bm25
sentence-transformers
python-dotenv
numpy
scikit-learn
chromadb==0.5.18
pydantic==2.10.1
pydantic-settings
python-multipart